Set loading screen before rendering content
In order to avoid bad redirections when reloading, set the
App state to loading at the beggining to make sure we have
session information before rendering.

The routes are being rendered before the session fetching
(async) finsihes, that leads to unauthorized access if we
already hav a session, because the backend didn't respond
yet, but we're trying to render a current route.