Add Api endpoint to process specification info
- The specification uploaded by the user may not
be valid, so this endpoint will be reached before
submitting the mapping upload form.

- This will be in charge of recognizing the domains
in the file. That's what concerns us right now but,
it's prepared to do other future necessary
validations.

Add prettier linter for js and jsx files

- Add a script to run prettier on pre commit git hook.
This is to ensure a unique coding style on the
frontend .