Filter Specification
- Build a specification from the one uploaded, but only
with the domain selected by the user, and the related
properties,  recursively.

Follow the steps shown in:
- [this diagram](https://miro.com/app/board/o9J_kna9728=/?moveToWidget=3074457349661692411&cot=6)

This is to ensure we show the user the preview of the
file containing only 1 domain. It may be the case of
the user uploading an already valid file. But it can
only be the case of the user uploading a file with
several domains, so always show the final version
(the valid one).