Add "edit" and "remove" Term functionality
Since the uri of the term is enough to identify the
related term, there's no need for this attribute to
be duplicated in both entities.

In order to maintain responsibilities separated
as much as possible, all the parsing operations
are now in a set of parsers per model.

The specs controller was handling lots of data.
It's better to keep things simple. So lets
separate the term and properties to be fetched
using an enpoint dedicated to it.

Edit term with only 2 attributes allowed to be updated
by now (already scheduled a meeting to discuss the rest
of the attributes).

Allow nested attributes on request for update terms with
its corresponding properties.

Remove a term inside the edit modal window.
Remove both from frontend and backend.

Implement raw preview on edit term

Use code-mirror library to make a textarea with
the raw json content of the term and property.
Make it view-only.

Fix file uploader to permit upload same file twice

When uploading a file, selecting a domain and then
clicking on "Re-Import", then trying to upload the
same file didn't work. The file input needed to be
reset.