Correctly draw the terms that was added to domains
In the mapping to domains view, after clicking on the "Done Domain Mapping"
button, the mapping terms are sucecssfully created, but the next page load,
the terms are again "unmapped".

Those specification "mapped" terms are the ones that are "in-memory" marked
as "mapped" and also those that are referenced by the current mapping terms
as "mapped_spec_terms".