Refactor the modeling of mapping, terms, and spine
The architecture schema was not prepared to represent multiple terms being
mapped to the same spine term. Different changes were performed in order to model that.

Add ExpandableOptions component

In order to model a custom select component to represent the predicates,
a new abstract component is proposed as an "abstract options" expand element.

This will serve firstly for the predicates, but also for the domain and
range of the terms, as a way to select one from multiple string options.

This component receives options as an array of javascript objects with a
name and id and can also receive 2 callbacks:
a. onClose
b. onExpand

Unset files if error on upload specification