Finish align and Fine Tune view
1. Prepare comment on alignment options as dropdown

2. Design of modal window for edit an alignment

3. Implement 'Comment on Alignment' view

4. Save the comment on the DB, and fetch it before render.

5. Add a 'collapsible' component

6. Save the changes on alignments to DB
Iterate thorugh all the alignments to save the changes to DB.
Make the button to save available only if not loading and if
there are any changes performed.
Save the mappedTerms as a many to many relation

7. Add synthetic element on UI & Save synthetic element added to the spine
This process consts of saving both:

The spine term, which is a property in the specification marked
as spine for the current domain, and
The alignment (mappingTerm) which specifies which mappedTerm/s
from the uploaded specification is going to be aligned with the
new synthetic term and with which predicate.
Make the button to save available only if not loading and if there are any changes performed.

8. Add secure error management for synthetic element

9. Save comment and predicate udpates for alignments

10. Add orange dot on an alignment with a comment

11. Add ability to change from edit to comment on edit alignment form

12. Correct comments on the PR
Filter by "changed" alignments before save, so we don't
save all the alginments every time.
Add transaction to handle alignment updates.
Improve error handling.

13. Implement outside alerter component
In order to shrink expandable components when the user clicks
outside its, we need to capture that event.
Implement it on 'Expandable' and on 'Collapsible'.

14. Improve Collapsible and Expandable components
Make it more configurable. Implement TermCard with Collapsible

MORE:
- Remove memory leak issues on Collapsible and ExpandableOptions components
- Simplify the update mapped terms logic
- Replace "Animated" component: The Animated component is causing problems on
"ComponentWillReceiveProps". That method is deprecated in React. Use instead,
"Styled Components" with "React Animations". Create a component to wrap all
animations.
- Propagate the name dropped on the left (property name) on Synthetic element
added to the spine.
- Permit only 1 term to drag when adding synthetic