Permit to revert selecting terms on mapping to domains screen
Once a user selects a term by dropping it into the domain box, there is no way
to undo it.

- Add an "x" icon to allow the user to revert term being selected.
- Implement the logic to delete a selected term from a mapping.
- Connect the UI with this new edpoint.
- Update the buttons availability logic.

Add ability to remove term from being mapped in Aling and Fine Tune screen

Once a user drops a term into a spine term in the Align and Fine Tune view, there's no
option to remove it from being mapped.

Permit to finish a mapping

- One its complete permit to change the status of the mapping to "mapped", in order
to mark this as completed.
- Only do it when all mapping terms are mapped to a spine term.