Refactor to integrate format converter
- Refactor term - specifications relation architecture
The specifications and terms relationship was meant to be one-to-many. This is not correct,
we need to represent terms to be part of one or many specifications.
It changed now so we improve the data handling in DB, impacting performance positively.

- Merge vocabularies when uploading multiple files
When the user uploads multiple files, before identifying domains (rdfs:classes),
we use the service to merge the files into one big JSON-LD file.

- Separate vocabularies on filtering, before the preview
With the purpose of being able to show the vocabularies on further screens, separate the vocabularies
as collections when filtering.

- Show the separated vocabularies on tabs
Work with the separated vocabularies obtained form the filtering phase, to
show it on tabs on the preview.

- Prepare a hook to integrate the "conversion of formats" work