Create API endpoint to retrieve audits
The audits can be fetched by using the api service. And also filtering by:
- class_name: Since the list might be a hugh list, we initially filter by
class name and ids, If the consumer needs more than one class, it can
make separate requests.

- audit_action: It might be the case when the consumer only needs "update"
or "create" audits.

- date_from: The consumer can specify "date_from" param to only retrieve
audits from that date on.

The order of the audits will be in ascendent way, meaning the newest first.