Add overall design for the "Synthetic Spine Property Mapping By Category" view.
Now we can see the mapping properties list in a separated route in the application.
This is just the design of the view. It fetches some data in order to accomplish
the correct view.

Details
Add design for each of the alignments of a spine term
Enable endpoint to retrieve all alignments for a spine term
List all alignments for a spine term.
Fetch these alignments from UI
Add design for property cards.
Handle to fetch the properties of the selected domain spine every time the
component updates, and also when the component mounts.
Add different styles for filters when are being shown completely and when
it's shown by a mouse hover
Add option to select/deselect all predicates
Select/Deselect Alignment options
Complete the UI functionality for selecting and deselecting:
Spine Organizations
Alignment Organizations
Alignments
on filters.
Fill the filters with options
Filters are now filled with organizations and predicates using the API
service to fetch.
Add design for "Hoverable Label"
In order to accomplish showing/hiding the options, we added a new component
that by default, shows a label that receives as a prop, and when moving the
mouse over it, will show content that also receives as a prop.
Add first filters row design
Add searchbar design
Design domain tabs
In the "Synthetic Spine Property Mapping By Category" view:

Fetch domains.
Build an abstract DESM tabs component.
Build the main component.