Support reading files with different structure
Prior file ingest logic:
- We expect a file to have a context and a graph
of nodes with different types.

Current file ingest logic:
- Now we can process files with or without those
keys. If a file does not contain a graph we review it to be the case of a
file with an array of nodes (an implicit graph).
- Now we also take care of the files without a context but, with the keys
inside the nodes being full URI's.

Prepare the backend to accept a list of domains to filter a specification

In the current workflow, at upload phase, we ask the user to select only one domain
to filter the file. Now we support filtering using multiple domains.

Support selecting more than 1 domain at upload phase

When a user uploads a file with more than 1 class (rdf:Class), we previously
ask the user to pick only 1 domain. This is an incovenience, since some domains
are defined by more than 1 rdf:Class.

Now we ask the user to select 1 or more domains, letting him search, as previously,
and also showing the user the amount of domains currently veing selected. Then we
filter the file using those domains the user selected, and let the user review.

This leads us to the generation of a more complete file.