Minor UI Improvements
- Implement removal of predicate when removing a term from being mapped
Once a user removes the last mapped term in a mapping, remove also the predicate choice.

- Remove synthetic term and alignment when the user removes all mapped terms in AFV.

- Design a new import specification button
With the purpose of being more self-explanatory to the user, the form submit
button with the text "Import Specification", now is not visible until there's
a file and appears as a circle button with no text at the bottom of the form.
It also contains a tooltip to guide the user on the action.

- Make Use Case and Version not required
In the upload mapping form, change 2 text inputs to not be required.
a- Use Case
b- Version

- Add email to organizations
Every organization must have an email. It will be necessary when implementing
broadcast notifications on mapping updates.

- Corrections on the PR from comments
a- Replacement of boolean condition variable usage. Since the result of the expression gets
rendered eventually, we need it to evaluate to a boolean value. Otherwise 0 will be displayed
when there are no files.
b- No need in proc: before_destroy on mapped_term model. Simply call the fucntion by it's
symbol name.