- Dynamically send the user from the specs list to the mappings list to the specific domain of the mapping Clicking the view button in the specifications list takes the user to the property alignments list but with the filters prepared to see only the alignments related to the selected mapping.
- Complete fixed alignment and spine terms ordering options
The options to order the alignments and spine terms, were not complete. We now show the options.
It's still lacking the specific ordering functionality.

- Remove all references to properties. Fix domain fixed reference.
There should be no interface references to "element" or "elements"...all should be to "property"
or "properties" since JSON-LD has properties, not elements. Same for RDF, properties, not
elements.

-Implement filtering by spine organizations and alignment organizations

- Implement predicates filter

- Permit to edit the scheme attribute in terms

- Correct wrong abstract class listed in elements selected box, fixed.

- Add scheme to properties
In order to properly print the scheme for a property, send it from the UI and save it into our
records using the service. The value is the name of the first uploaded file.

- Implement hide spine terms with no alignments
The checkbox functionality to hide spine terms that have no results (meaning alignments to it)
in the "Synthetic Spine Property Mapping By Category" view.

- Filter spine terms by the value entered in the search box