Add spine terms and alignment order options
In order to let the user pick a sorting criteria when listing spine properties and
alignments to each of it, we list the options and let the components interact to
know in each operation which the value of these sort options is. It's still lacking
the use of these values to properly sort the results.