Implement strong password validation
When a user is added and access the platform through their
email link to sign in/reset password, the new password to
be generated is now validated to be strong.

In this validation, we use an entropy based analysis. It is,
checking the password deductibility, configuring it first to
have 18 as entropy level, which is acceptable as a security
level.

Secure password inclusion in request payloads

Use JSON Web Token both to encode in the frontend and to decode
in the service to make the password secure when passsing it as
a parameter inside the request payload in both:
- Check user password strength
- Reset user password

Use environment variables to make the algorithm and secret key to
encode/decode JWT not traceable.