Infer node types by following definitions
Improve user experience on uploading specs with failures

- Warn the user to select domain if not selected
- Fail with a toast error message if not classes found.
- Include JSON Schema converter to JSON-LD

Build JsonLd parser to infer node types

In Json-ld the type of a node, may be expressed in different ways. It could
be a full HTTP URI, an prefixed URI, or it may not be typed. In order to
recognize classes and properties, it may be necessary to fetch resources in
the internet.

Now we can do that for each node. It implies a performance impact, but it's
planned to incorporate caching before it's merged, because many nodes of the
now same type are being analyzed repeatedly in separate requests, because
we don't have chacing at the moment.

Refactor specifications processor

To be consistent with the new JsonLd parser, make the processor
not static and place methods in the proper classes.

Implement cache using redis to save node types

Many node types might be repeated, so we cache them in order to minimize
the HTTP requests.

Adjust the specifications processor to create spec properly

After the latest changes in the jsonld node parser and the new ways
of recognizing classes and properties, the processor was not creating
the same specification as in the preview.

Add validations

- Consider comparing with shrinked and expanded uri when searching for properties
of a given class.

- Validate prefixes are in place inside the provided context before asccesing it.

- Validate specifications format on initialize of Specifications parser.

Add PESC, CEDS and RDF ASN compatibility with new changes