Improve the user experience with textual messages
Use tooltips on buttons and action options to let the user know about the behaviour
and course of action of each stage of the application.

Do not allow the specification to be created if the tool couldn't recognize at
least one property in the upload phase.

Fix specification creation error on upload phase.