Permit to change the status of a mapping
When a mapping is marked as "mapped" (finished), now we can decide to put it
back to "in-progress". Doing this will allow us to change a predicate, and/or
the terms of the specification.

In addition, we also can grab a mapping that's marked as "in-progress", and
put it back to "uploaded". This will show us all the original terms of the
uploaded specification, allowing us to change those terms (selecting/
deselecting).