Validate the schema of a configuration profile using json schema validator gem
- Try json_schema_tools (more complicated to use)
- Try json-schemmer (1 additional step)
- Install json-schema gem
- Build 2 schemas:
  - The first to validate JSON structures and determine if it represents a valid Configuration Profile.
  - The second, to validate if a JSON structure is valid (although not complete) Configuration Profile.
- Add tests

Move the JSON schema validation logic to the configuration profile model

In order to get the CP to know if its structure is valid or not, and having the logic to be not too extense, the
decision was to move all the logic to the model.

Implement validation with real complete and valid JSON Schemas for configuration profiles.

In order to validate the configuration profile structure against the JSON schemas that defines valid and complete objects:
- Define a JSON schema for valid configuration profiles.
- Define a JSON schema for complete configuration profiles.
- Make the configuration profile able to determine whether it's valid or not by using a JSON Schema validator against its own structure.
- Redirect the configuration profile state logic to make the transition to complete by using this new configuration profile method.

Raise error if trying to complete with incomplete structure in configuration profiles