Feature/issue97 generate all cp related models when activating (#99)
* Generate interactor for creating Agents

An agent is understod as a user with a role in the DESM platform. So this interactor will receive the
user information along with the role and create it.

Create the Mapping Predicates generator interactor

- Move the logic to consume skos file from a URI to a module so we can leverage it multiple times.
- Create the interactor and tests.

Create concept schemes interactor

- Following the line of the previous interactors, create one for the concept schemes.

Add create schema interactor

Allow to create a schema with its name, domain, user, and a uri to fetch from.
Generate its terms.

Create interactor that generates all the necessary structure for a configuration profile

- Trigger when activating the profile
- Create each entity based on parsing the json data representing the structure
- Still need to parse the schemes and concept schemes

Create the "configuration profile structure generator" interactor

- This interactor will use all the previously created interactors and
- Follow the JSON "structure" attribute of the completed cofiguration profile and create:
  - The administrator for the configuration profile
  - The abstract classes
  - The mapping predicates
  - The DSO's and all the related information about it (administrator, schemas, concept schemes)
- Make it work inside a transaction
- Tests

* Remove scripts and update readme "Collaborate" section

Now that we have linting and testing validation instances in a GitHub Actions workflow for every
Pull Request, we don't need the pre-commit and pre-push git hooks.