Show progress to the user while activating a configuration profile (#110)
* Show progress to the user while activating a configuration profile

To activate a configuration profile, if its state is 'complete', implies creating all the related data,
like DSO's, agents, schemes and more. So depending on the amount of data specified, it may take a few minutes.
Now we show a progress locked window to the user when performing this task.