Design Dso info tabs
The Dso Info step of the edit configuration profile view has more information than the others.
It not only contains metadata about the configuration profile, but also includes agents and schema
files and concept schemes information.

As in the mockups, the information is presented in this view in a tab system. This commit includes
the metadata view as the first tab and prepares to add the rest of the tabs in a future commit.