Change the validation of the structure of a configuration profile
The structure of a configuration profile is now being processed as JSON but, in different stages,
it will have different variable naming conventions, as follows:
- When the frontend manipulates the structure information, it will do it in camelcase.
- When the frontend sends the structure information to the backend, it will first transform
itto snakecase.
- When the backend receives the structure information, it will expect snakecase.
- The structure information will be saved into the database in snake_case.
- The "complete" and "valid" schemas are stored in the repository as json files using camelcase,
so when the backend uses it to validate the information saved, it will transform the schemas to
snakecase.
- When the frontend requires the structure information from the backend, this last will send it
just as it is, but when it comes into the frontend, it will transform it into camelcase.